requirements.txt
pyicloud
geopy
python-geoip
python-geoip-geolite2
win_inet_pton